Reverted wrapper changes
